gateway_service/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
httpx==0.24.0
grpcio-tools==1.71.0
grpcio==1.71.0
googleapis-common-protos==1.69.2
PyJWT
